LayoutTests/fast/events/touch/ios/programmatic-scrolling-on-touchmove.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->
<html>
<meta name="viewport" content="width=device-width">
<head>
<script src="../../../../resources/js-test.js"></script>
<script src="../../../../resources/ui-helper.js"></script>
<style>
html, body {
    margin: 0;
    text-align: center;
    width: 100%;
    font-family: system-ui;
}

.container {
    background: linear-gradient(to bottom, red 0%, green 50%, blue 100%);
    width: 100%;
    height: 3000px;
    display: inline-block;
}
</style>
<script>
jsTestIsAsync = true;

addEventListener("load", async () => {
    description("To manually run this test, swipe up to scroll down; scrolling should not be jittery.");
    let scrollStartTimerID = null;
    let startPosition = 0;
    let scrollDeltas = [];
    const container = document.querySelector(".container");
    container.addEventListener("touchstart", (event) => {
        startPosition = event.touches[0].pageY;
        event.preventDefault();
    });

    container.addEventListener("touchmove", (event) => {
        const position = Math.round(event.touches[0].pageY);
        scrollTo(0, pageYOffset + startPosition - position);
        event.preventDefault();
    });

    addEventListener("scroll", () => {
        if (pageYOffset < 0 || pageYOffset > 1000)
            testFailed(`Unexpected pageYOffset: ${pageYOffset}`);
    });

    await UIHelper.sendEventStream(new UIHelper.EventStreamBuilder()
        .begin(100, 450)
        .move(100, 50, 0.5)
        .end()
        .takeResult());

    shouldBeGreaterThan("pageYOffset", "0");
    finishJSTest();
});
</script>
</head>
<body>
    <p id="description"></p>
    <div class="container"></div>
</body>
</html>
